coord -> int to double

+++ include/Physics.hpp
@@ -20,17 +20,17 @@ public:
   /*                                  */
   /************************************/
 
-  Physics(const Vect<2, int>& planetCenter, double radius, double mass) : _planet(planetCenter, Vect<2, int>(0, 0), radius, mass) {}
+  Physics(const Vect<2, double>& planetCenter, double radius, double mass) : _planet(planetCenter, Vect<2, double>(0, 0), radius, mass) {}
 
   struct                Fixture
   {
-    Fixture(const Vect<2, int>& i_pos, const Vect<2, int>& i_speed, double i_radius, double i_mass) : pos(i_pos),
+    Fixture(const Vect<2, double>& i_pos, const Vect<2, double>& i_speed, double i_radius, double i_mass) : pos(i_pos),
                                                                                                       speed(i_speed),
                                                                                                       radius(i_radius),
                                                                                                       mass(i_mass){}
 
-    Vect<2, int>        pos;
-    Vect<2, int>        speed;
+    Vect<2, double>     pos;
+    Vect<2, double>     speed;
     double              radius;
     double              mass;
   };
